Add functional requirements
