+++ src/main/java/fa/training/dao/OrderDAO.java
@@ -12,117 +12,9 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class OrderDAO {
-    Database database;
-    public OrderDAO(){
-        database = new Database();
-    }
-
-    public int create(Order order) throws SQLException {
-        Connection conn = database.getConn();
-        int count;
-
-        try{
-            conn.setAutoCommit(false);
-            String sql = "INSERT INTO `orders`(`order_date`, `customer_id`, `order_id`, `total`) VALUES ('?',?,?,?)";
-            PreparedStatement preparedStatement = database.getConn().prepareStatement(sql);
-            preparedStatement.setDate(1, (java.sql.Date)order.getOrderDate());
-            preparedStatement.setInt(2, order.getCustomerId());
-            preparedStatement.setInt(3, order.getOrderId());
-            preparedStatement.setDouble(4, order.getTotal());
-            count = preparedStatement.executeUpdate();
-            conn.commit();
-            conn.setAutoCommit(true);
-
-            return count;
-        }catch (SQLException e){
-            conn.rollback();
-
-            return 0;
-        }
-    }
-
-    public List<Order> retrieve(){
-        List<Order> list = new ArrayList<>();
-        Order order;
-
-        try{
-            String sql = "SELECT * FROM `Orders`";
-            Statement statement = database.getConn().createStatement();
-            ResultSet resultSet = statement.executeQuery(sql);
-            while (resultSet.next()){
-                order = new Order(resultSet.getInt("order_id"),
-                        (resultSet.getDate("order_date")),
-                        resultSet.getInt("customer_id"),
-                        resultSet.getInt("employee_id"),
-                        resultSet.getDouble("total"));
-                list.add(order);
-            }
-        }
-        catch (SQLException e){
-            throw new RuntimeException(e);
-        }
-        return list;
-    }
-
-    private int updateDate(int order_id, int customer_id, int employee_id, double total) throws SQLException{
-        Connection conn = database.getConn();
-        int count = 0;
-        
-        try{
-            conn.setAutoCommit(false);
-
-            if((employee_id < 0) && (total < 0)){
-                String sql = "UPDATE `Orders` SET `customer_id` = ? WHERE `order_id` = ?";
-                PreparedStatement preparedStatement = database.getConn().prepareStatement(sql);
-                preparedStatement.setInt(1, customer_id);
-                preparedStatement.setInt(2,order_id);
-                count = preparedStatement.executeUpdate();
-            }
-
-            if((customer_id < 0) && (total < 0)){
-                String sql = "UPDATE `Orders` SET `employee` = ? WHERE `rrder_id` = ? ";
-                PreparedStatement preparedStatement = database.getConn().prepareStatement(sql);
-                preparedStatement.setInt(1,employee_id);
-                preparedStatement.setInt(2,order_id);
-                count = preparedStatement.executeUpdate();
-            }
-
-            if((customer_id < 0) && (employee_id < 0)){
-                String sql = "UPDATE `Orders` SET `total` = ? WHERE `order_id` = ? ";
-                PreparedStatement preparedStatement = database.getConn().prepareStatement(sql);
-                preparedStatement.setDouble(1, total);
-                preparedStatement.setInt(2,order_id);
-                count = preparedStatement.executeUpdate();
-            }
-
-            conn.commit();
-            conn.setAutoCommit(true);
-
-            return count;
-        }catch(SQLException e){
-            conn.rollback();
-            return count;
-        }
-    }
-    private int delete(int order_id) throws SQLException{
-        Connection conn = database.getConn();
-        int count;
-
-        try{
-            conn.setAutoCommit(false);
-            String sql = "DELETE FROM `Orders` WHERE `order_id` = ?";
-            PreparedStatement preparedStatement = database.getConn().prepareStatement(sql);
-            preparedStatement.setInt(1,order_id);
-            count = preparedStatement.executeUpdate();
-            conn.commit();
-            conn.setAutoCommit(true);
-
-            return count;
-        }catch(SQLException e){
-            conn.rollback();
-
-            return 0;
-        }
-    }
+interface OrderDAO {
+    public int create(Order order) throws SQLException;
+    public List<Order> retrieve();
+    public int updateDate(int order_id, int customer_id, int employee_id, double total) throws SQLException;
+    public int delete(int order_id) throws SQLException;
 }
